Rename compression package to compressors
