+++ docs/modules/ROOT/pages/compressors/gzip.adoc
@@ -1,0 +1,17 @@
+= gzip
+
+Here is a simple example of how to use gzip compression.
+
+== Example Usage
+
+.Gzip Compression
+[,java]
+----
+include::example$java/com/example/compressors/GzipExamples.java[tag=gzip-compressor,indent=0]
+----
+
+.Gzip Decompression
+[,java]
+----
+include::example$java/com/example/compressors/GzipExamples.java[tag=gzip-decompressor,indent=0]
+----
