+++ src/main/java/io/github/compress4j/compressors/Compressor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.compress4j.compression;
+package io.github.compress4j.compressors;
 
 import java.io.File;
 import java.io.IOException;
