+++ src/test/java/io/github/compress4j/compressors/bzip2/BZip2CompressorBuilderTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.compress4j.compression.bzip2;
+package io.github.compress4j.compressors.bzip2;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
