+++ src/test/java/io/github/compress4j/compressors/bzip2/BZip2CompressorTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.compress4j.compression.bzip2;
+package io.github.compress4j.compressors.bzip2;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
