+++ src/testFixtures/java/io/github/compress4j/compressors/memory/InMemoryCompressor.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.compress4j.compression.memory;
+package io.github.compress4j.compressors.memory;
 
-import io.github.compress4j.compression.Compressor;
+import io.github.compress4j.compressors.Compressor;
 import java.io.IOException;
 import java.io.OutputStream;
